+++ test/cases/loader-option-watchFiles/webpack.config.js
@@ -1,0 +1,31 @@
+const path = require('path');
+const HtmlBundlerPlugin = require('../../../');
+
+module.exports = {
+  mode: 'production',
+
+  output: {
+    path: path.join(__dirname, 'dist/'),
+    clean: true,
+  },
+
+  plugins: [
+    new HtmlBundlerPlugin({
+      entry: {
+        index: './src/index.html',
+      },
+    }),
+  ],
+
+  module: {
+    rules: [
+      {
+        test: /\.(html)$/,
+        loader: HtmlBundlerPlugin.loader,
+        options: {
+          watchFiles: /\.(s?css)$/i,
+        },
+      },
+    ],
+  },
+};